honeycomb setup (celery)
this pulls in the settings (will require
https://github.com/appsembler/edx-configs/pull/875).

It enables basic honeycomb tracing on celery tasks.

It does not configure the django middleware. I still need to do some
work to make a version of that that will work with
`MIDDLEWARE_CLASSES`. So we don't have tracing on gunicorn/HTTP
requests, but I realized that we could still get tracing on the celery
tasks, which still has value.