add honeycomb legacy middleware
the default honeycomb middleware in `beeline` requires the
newer (Django 1.10+) `MIDDLEWARE` setting and will not work with older
`MIDDLEWARE_CLASSES`, which is what Hawthorn still uses.

Until we are on a newer release that uses the newer API, this adds a
minimal honeycomb middleware to trace basic HTTP request/responses.

It doesn't do anything fancy like database query tracing (even the
beeline version needs Django 2.0+ for that) but it should get us the
basic fields in the same format and will enable us to add custom
traces of our own when we want to instrument particular parts of the
code.