honeycomb trace course home view
I'm trying to fill in the big blank spot in this trace:

https://ui.honeycomb.io/appsembler/datasets/tahoe-hawthorn-prod/result/AKJBe6fg7ra/trace/56WUo6hWzP7

for the course home view. Taking a bit of a shotgun approach since the
code uses `web_fragments.views.FragmentView` as an abstract base class
and the execution bounces around a bit so I'm not totally sure what is
getting called, but this should help me narrow that down.