Populate the registration email field, given invitation code
* Use case is for learners who are invited to a site and enroll in one
or more courses with the invitation
* The Tahoe course enrollment API can create an enrollment invitation,
which creates a record in `student.models.CourseEnrollmentAllowed`. This
requires the learner to enter the same email address used for the Tahoe
course enrollment API 'create' call.
* This commit uses the new `appsembler.invitations` package to find the
email used in the Tahoe course enrollment API. A system external to
Tahoe provides the invitation code to the learner in the Tahoe
registration link as the `activation-code` query parameter,find the
matching email address and fill the registration form with this email
address